### 2024-01-02 #6178 @editorialbot
[PRE REVIEW]: DFM Generator v2.2: a new method for simulating the growth of natural fracture networks
**Submitting author:** @JointFlow (<a href="http://orcid.org/0000-0001-7868-5079">Michael Welch</a>)
**Repository:** https://github.com/JointFlow/DFMGenerator
**Branch with paper.md** (empty if default branch):
**Version:** v2.2.2
