### 2024-01-09 f30fed8
Modified DFMGeneratorUI.cs to set the correct unit label for Geothermal Gradient on the Petrel UI
