### 2024-01-11 #6178 @kthyng
Previous pre-review issue: https://github.com/openjournals/joss-reviews/issues/5066
