### 2024-01-15 bd71195
Modifications to FractureDipSet.calculateTotalMacrofracturePopulation(), FractureDipSet.RecalculateElasticResponse() and Tensor2S.GetEigenvalues() to fix bugs B74, B80 and B81 respectively. Needs to be recompiled and tested.
