### 2024-01-22 fb4f180
Modified FractureDipSet.GetOptimalDuration() and GridblockConfiguration.CalculateFractureData() to automatically recalculate horizontal strain partitioning when the strain azimuth changes within a deformation episode, and to calculate time until driving stress reaches zero when it is decreasing. Also changed the error message when the gridblock timestep limit is reached. Not recompiled.
