+++ DFMGenerator_SharedCode/FractureDipSet.cs
@@ -2633,81 +2633,90 @@ namespace DFMGenerator_SharedCode
 
             // Now we can calculate the optimal timestep duration for this fracture set
 
-            // If the fracture set has been deactivated, do not calculate an optimal duration
-            // Instead we return the default value infinity (no optimal duration calculated)
-            if (CurrentFractureData.EvolutionStage == FractureEvolutionStage.Deactivated)
-            {
-                // No calculation required
-            }
-            // If the initial driving stress is negative, the optimal timestep duration should be the time until it becomes positive
+            // First we will calculate the time taken for the driving stress to reach zero (from either a positive or negative value) based on the rate of change of driving stress
+            // This will act as an upper bound to the timestep length
+            // If the driving stress is less than zero it will represent the time until the fracture set becomes active
+            // If the driving stress is greater than zero it will represent the time until the fracture set stops propagating
+            // This can only be calculated if the initial driving stress U is not zero and the rate of change of driving stress V is in the opposite direction to U
             // In fact we compare the initial driving stress to the maximum rounding error, not zero, to determine if this is the case.
             // This is because calculating the time required for driving stress to reach zero, multiplying it by horizontal strain rate to increment horizontal strain, and using new horizontal strain to calculate driving stress does not always give a driving stress = 0 (as it should).
             // Sometimes due to rounding errors in the calculation, it generates a slightly negative driving stress. This can cause the calculation to get stuck in a loop, with no strain increments and no fracture growth, until the maximum number of timesteps is reached.
-            else if (U < -PreviousFractureData.MaxDrivingStressRoundingError)
+            if ((Math.Abs(U) > PreviousFractureData.MaxDrivingStressRoundingError) && (Math.Sign(U) != Math.Sign(V)) && ((float)Math.Abs(V) > 0f))
             {
-                // First we can calculate the time taken for the driving stress to reach zero based on the rate of change of driving stress
-                // Note that we must be dealing with Mode 2 shear fractures at this point, because if the fractures were dilatant then U must already be positive or zero
-                // Therefore this represents the time until Mode 2 shear displacement will occur   
-                // If the driving stress is static or decreasing then it will never reach zero, so we cannot calculate an optimal duration in this way
-                if (V > 0)
-                {
-                    // However calculating the time taken for the driving stress to reach zero for a Mode 2 fracture is more complicated than simply dividing the negative initial driving stress -U by the rate of change of driving stress V
-                    // As we have noted previously, the rate of change of driving stress may itself change through time, as the shear displacement vector and in situ stress change
-                    // We must therefore use a quadratic expression comprising the three orthogonal components of the stress acting on the fault: normal, shear in the direction of strike, and shear in the downdip direction
-                    // The rates of change of these components do not change through time
+                // Calculating the time taken for the driving stress to reach zero for a fracture with shear displacement is more complicated than simply dividing the negative initial driving stress -U by the rate of change of driving stress V
+                // As we have noted previously, the rate of change of driving stress may itself change through time, as the shear displacement vector and in situ stress change
+                // We must therefore use a quadratic expression comprising the three orthogonal components of the stress acting on the fault: normal, shear in the direction of strike, and shear in the downdip direction
+                // The rates of change of these components do not change through time
 
[…]
