### 2024-01-22 b69a0d7
Removed Excel output files from Test Models folder, so Standalone zip file is under the Github limit
Recompiled Pip installer files and standalone code. Need to test installers and run test models before ready for release

Modified FractureDipSet so it does not revert from ResidualActive to Growing within the calculateTotalMacrofracturePopulation() function, but instead within the fds.checkStressShadowWidthChange() function, only if the stress shadow width has changed. This fixes the problem (B75) but still needs testing on multi-stage models

A number of Petrel UI enhancements, including: 1) 
[…]
