+++ DFMGenerator_SharedCode/FractureDipSet.cs
@@ -76,6 +76,14 @@ namespace DFMGenerator_SharedCode
         /// </summary>
         private double dip;
         /// <summary>
+        /// Variable for sine of fracture dip
+        /// </summary>
+        private double sindip;
+        /// <summary>
+        /// Variable for cosine of fracture dip
+        /// </summary>
+        private double cosdip;
+        /// <summary>
         /// Unit vector on the fracture plane in direction of maximum dip (in the case of bimodal conjugate dipsets, dipping in J+ direction)
         /// </summary>
         private VectorXYZ dipVector;
@@ -97,6 +105,10 @@ namespace DFMGenerator_SharedCode
             {
                 // Set the fracture dip
                 dip = value;
+
+                // Set the sine and cosine of fracture dip
+                sindip = VectorXYZ.Sin_trim(Dip);
+                cosdip = VectorXYZ.Cos_trim(Dip);
 
                 // Create a new normal vector
                 normalVector = VectorXYZ.GetNormalToPlane(fs.Strike + (Math.PI / 2), value);
@@ -1329,7 +1341,7 @@ namespace DFMGenerator_SharedCode
         /// <summary>
         /// Pitch of the shear stress vector on the surface of the fractures relative to fracture strike (radians, positive downwards; will return NaN if shear stress is zero)
         /// </summary>
-        public double ShearStressPitch{ get; private set; }
+        public double ShearStressPitch { get; private set; }
         /// <summary>
         /// Pitch of the shear displacement vector on the surface of the fractures relative to fracture strike (radians, positive downwards; will return NaN if shear displacement is zero)
         /// </summary>
@@ -1491,33 +1503,18 @@ namespace DFMGenerator_SharedCode
                             // Recalculate fracture mode factors
                             if (BiazimuthalConjugate)
                             {
-                                double sindip = Math.Sin(Dip);
-                                double Maa = sindip;
-                                double Mas = 0;
-                                double Mss = sindip / 2;
-                                Maa_eaa2d_eh2d = fs.eaa2d_eh2d * Maa;
-                                Mas_eaaasd_eh2d = fs.eaaasd_eh2d * Mas;
-                                Mss_eas2d_eh2d = fs.eas2d_eh2d * Mss;
+                                Maa = sindip;
+                                Mas = 0;
+                                Mss = sindip / 2;
                             }
                             else
                             {
-                                double sindip = Math.Sin(Dip);
                                 double oneMinus2Nur = 1 - (2 * gbc.MechProps.Nu_r);
-                                double Mff = Math.Pow(oneMinusNur, 2) / oneMinus2Nur;
-                                double Mfw = 0;
-                                double Mww = oneMinusNur / 2;
-                                double Mfs = 0;
-                                double Mss = 1 / 2;
-                                Maa_eaa2d_eh2d = ((eff2d_e2d * Mff) + (efffwd_e2d * Mfw) + (efw2d_e2d * Mww)) / sindip;
-                                Mas_eaaasd_eh2d = (efffsd_e2d * Mfs) / sindip;
-                                Mss_eas2d_eh2d = (efs2d_e2d * Mss) / sindip;
-                            }
-                            // Remove rounding errors
-                            {
-                                double Mtot = Maa_eaa2d_eh2d + Mas_eaaasd_eh2d + Mss_eas2d_eh2d;
-                                if ((float)(Mtot + Maa_eaa2d_eh2d) == (float)Mtot) Maa_eaa2d_eh2d = 0;
-                                if ((float)(Mtot + Mas_eaaasd_eh2d) == (float)Mtot) Mas_eaaasd_eh2d = 0;
[…]
