+++ DFMGenerator_SharedCode/GridblockConfiguration.cs
@@ -23,6 +23,14 @@ namespace DFMGenerator_SharedCode
     /// Enumerator for global grid directions (N = Y+, E = X+, S = Y-, W = X-)
     /// </summary>
     public enum GridDirection { N, E, S, W, None }
+    /// <summary>
+    /// Enumerator for return codes for the CalculateFractureData function: 0 if the calculation runs to completion without errors; 1 if the timestep limit is hit
+    /// </summary>
+    public enum CalculateFractureDataReturnCode { Completed, TimestepLimitExceeded }
+    /// <summary>
+    /// Enumerator for return codes for the PropagateDFN function: 0 if the calculation runs to completion without errors; 1 if the gridblock geometry is not correctly defined; 2 if there is an error in the driving stress or propagation distance; 3 if fracture limit is hit
+    /// </summary>
+    public enum PropagateDFNReturnCode { Completed, GridblockGeometryError, DrivingStressError, NewFractureLimitExceeded }
 
     /// <summary>
     /// Class representing an entire gridblock
@@ -2063,8 +2071,8 @@ namespace DFMGenerator_SharedCode
         /// <summary>
         /// Calculate the implicit fracture model based on the parameters specified in the existing GridblockConfiguration.PropagationControl object
         /// </summary>
-        /// <returns>True if the calculation runs to completion before hitting the timestep limit; false if the timestep limit is hit</returns>
-        public bool CalculateFractureData()
+        /// <returns>CalculateFractureDataReturnCode object indicating if the calculation ran to completion or errors were encountered</returns>
+        public CalculateFractureDataReturnCode CalculateFractureData()
         {
             // Declare local variables
             bool CalculationCompleted = false;
@@ -2145,10 +2153,12 @@ namespace DFMGenerator_SharedCode
                 string FSheader2 = "Fracture stage\tDriving stress\tDisplacement sense\tSlip pitch\ta_uFP30\ts_uFP30\ta_uFP32\ts_uFP32\ta_MFP30\tsII_MFP30\tsIJ_MFP30\ta_MFP32\ts_MFP32\tClear zone volume\t";
 #if LOGDFNPOP
                 headerLine1 = string.Format("Timestep\tDuration ({0})\tEnd Time ({0})\t{1}\t{2}\t{3}\t{4}\t", timeUnits, "Sigma_eff.XX", "Sigma_eff.YY", "Sigma_eff.XY", "Sigma_eff.ZZ");
-                FSheader2 = string.Format("{0}\t{1}\t{2}\t{3}\t{4}\t{5}\t{6}\t{7}\t{8}\t{9}\t{10}\t{11}\t{12}\t{13}\t", "getEvolutionStage", "getFinalDrivingStressSigmaD", "Mode", "Mean_Azimuthal_MF_StressShadowWidth", "Mean_Shear_MF_StressShadowWidth", "Mean_MF_StressShadowWidth", "getInverseStressShadowVolume", "getInverseStressShadowVolumeAllFS",
-                    "getClearZoneVolume", "sII_MFP30_total", "sIJ_MFP30_total", "a_MFP32_total", "s_MFP32_total", "getClearZoneVolumeAllFS");
-                //FSheader2 = string.Format("{0}\t{1}\t{2}\t{3}\t{4}\t{5}\t{6}\t{7}\t{8}\t{9}\t{10}\t{11}\t{12}\t{13}\t", "getEvolutionStage", "getFinalDrivingStressSigmaD", "Mode", "getPhi", "getInstantaneousF", "getMeanMFPropagationRate", "getConstantDrivingStressU", "getVariableDrivingStressV", 
-                //    "a_MFP30_total", "sII_MFP30_total", "sIJ_MFP30_total", "a_MFP32_total", "s_MFP32_total", "getClearZoneVolume");
[…]
