### 2024-02-21 36651f3
Modified error handling when generating implicit fracture data and explicit DFNs to only output error messages at the end of the model run, not at every intermediate stage. Needs testing.
