+++ DFMGenerator_Standalone/Program.cs
@@ -166,7 +166,7 @@ namespace DFMGenerator_Standalone
                 input_file.WriteLine("% Output DFM at intermediate stages of fracture growth");
                 input_file.WriteLine("NoIntermediateOutputs 0");
                 input_file.WriteLine("% Flag to control interval between output of intermediate stage DFMs:");
-                input_file.WriteLine("%      - EqualArea (output at at approximately regular intervals of total fracture area)");
+                input_file.WriteLine("%      - EqualArea (output at at approximately euqal increments of total fracture area)");
                 input_file.WriteLine("%      - EqualTime (output at equal intervals of time)");
                 input_file.WriteLine("%      - SpecifiedTime (output at the end of each specified deformation episode)");
                 input_file.WriteLine("IntermediateOutputIntervalControl EqualArea");
@@ -629,7 +629,10 @@ namespace DFMGenerator_Standalone
             DFNFileType OutputDFNFileType = DFNFileType.ASCII;
             // Output DFM at intermediate stages of fracture growth
             int NoIntermediateOutputs = 0;
-            // Flag to control interval between output of intermediate stage DFMs; they can either be output at specified times, at equal intervals of time, or at approximately regular intervals of total fracture area
+            // Flag to control interval between output of intermediate stage DFMs:
+            // - EqualArea (output at at approximately euqal increments of total fracture area)
+            // - EqualTime (output at equal intervals of time)
+            // - SpecifiedTime (output at the end of each specified deformation episode)
             IntermediateOutputInterval IntermediateOutputIntervalControl = IntermediateOutputInterval.EqualArea;
             // Flag to output the macrofracture centrepoints as a polyline, in addition to the macrofracture cornerpoints
             bool OutputCentrepoints = false;
