### 2024-02-29 157a924
Rewritten MacrofractureXYZ.PopulateData() to bevel segment corners only if the I coordinate of the crossover point between the edges of adjacent segments lies between the I coordinate of the two segment cornerpoints, rather than using a minimum angle cutoff (B84), but not yet tested. Modified FractureSet.checkStressShadowInteraction() to not create relay zones if a third fracture lies between the two fracture tips (B85), but not yet working.
