+++ DFMGenerator_SharedCode/DFNComponents.cs
@@ -570,7 +570,11 @@ namespace DFMGenerator_SharedCode
         /// </summary>
         public GridDirection TrackingBoundary { get { return trackingboundary; } set { if (NonPropNodeBoundary == value) { trackingboundary = value; propnodeboundary = value; } } }
         /// <summary>
-        /// Direction of fracture propagation in local IJK coordinate frame - this will not change after fracture segment is initiated
+        /// Direction of the outer node relative to the inner node in the local IJK coordinate frame
+        /// </summary>
+        public PropagationDirection LocalOrientation { get { if (!reverseNodes) return LocalPropDir; else if (LocalPropDir == PropagationDirection.IPlus) return PropagationDirection.IMinus; else return PropagationDirection.IPlus; } }
+        /// <summary>
+        /// Direction of fracture propagation in the local IJK coordinate frame - this will not change after fracture segment is initiated
         /// </summary>
         public PropagationDirection LocalPropDir { get; private set; }
         /// <summary>
@@ -1134,7 +1138,7 @@ namespace DFMGenerator_SharedCode
         /// <summary>
         /// Links to contiguous macrofracture segments in the local (gridblock) DFN propagating in the IPlus and IMinus directions
         /// </summary>
-        private Dictionary<PropagationDirection, List<MacrofractureSegmentIJK>> MF_segments;
+        public Dictionary<PropagationDirection, List<MacrofractureSegmentIJK>> MF_segments;
         /// <summary>
         /// Add a local macrofracture segment to this fracture
         /// </summary>
@@ -1341,6 +1345,11 @@ namespace DFMGenerator_SharedCode
             SegmentCompressibility[PropagationDirection.IPlus].Clear();
             SegmentCompressibility[PropagationDirection.IMinus].Clear();
 
+            // Create two arrays of flags for whether the upper and lower segment cornerpoints will need later adjustment
+            // Segment cornerpoints will need adjusting they lie outside the cornerpoints of the adjacent segments, due to bevelling
+            Dictionary<PropagationDirection, List<bool>> adjustUpperInnerCornerPoints = new Dictionary<PropagationDirection, List<bool>>();
+            Dictionary<PropagationDirection, List<bool>> adjustLowerInnerCornerPoints = new Dictionary<PropagationDirection, List<bool>>();
+
             // Create a new list of cornerpoints for each segment, used to recreate the fracture as a series of contiguous planar segments
             // Move along the IPlus and IMinus sides of the fracture in turn
             foreach (PropagationDirection dir in Enum.GetValues(typeof(PropagationDirection)).Cast<PropagationDirection>())
@@ -1362,11 +1371,11 @@ namespace DFMGenerator_SharedCode
                 PointXYZ ThisSegment_UpperOuterBevelledCornerPoint = null;
                 PointXYZ ThisSegment_LowerOuterBevelledCornerPoint = null;
 
-                // Create two arrays of flags for whether the upper and lower segment cornerpoints will need later adjustment
-                // Segment cornerpoints will need adjusting if we cannot bevel them because the adjacent upper or lower boundaries are near parallel (as determined by the minimum bevelling angle)
-                // In this case it is possible they may lie outside the cornerpoints of the adjacent segments, so we will adjust them to lie between the cornerpoints of the adjacent segments
-                List<bool> adjustUpperOuterCornerPoints = new List<bool>();
-                List<bool> adjustLowerOuterCornerPoints = new List<bool>();
+                // Create arrays of flags for whether the upper and lower segment cornerpoints will need later adjustment, and add them to the appropriate dictionaries
+                List<bool> adjustUpperInnerCornerPointsCurrentSide = new List<bool>();
+                List<bool> adjustLowerInnerCornerPointsCurrentSide = new List<bool>();
[…]
