### 2024-03-05 0df2d56
Created a list of XYZ centrepoints within the MacrofractureXYZ object and populate this in the PopulateData() function, as the fracture patch cornerpoints are; since this is called at the end of each stage, it will populate it with the centrepoint locations at that time, so there will be no mismatch between the centrepoint and segments of the early stages of the DFN if individual segments continue to grow, as occurs if the centrepoints simply reference still active MacrofractureSegmentIJK nodes (B86).
