+++ DFMGenerator_Ocean/DFMGeneratorPlugin.cs
@@ -9,7 +9,7 @@ namespace DFMGenerator_Ocean
     {
         public override string AppVersion
         {
-            get { return "2020.1"; }
+            get { return "2023.1"; }
         }
 
         public override string Author
