### 2024-03-07 e678ae2
Modified FractureDipSet.Mean_MF_HalfLength() to check for MFP30=0 and if so returning zero (B87). Not yet tested.
