### 2024-03-09 910c3ba
Modified FractureSet.Calculate_MeanPropagationDistance() so that when it checks the mean propagation rate is greater than zero, it converts to a float to eliminate rounding errors; this prevents it from returning infinite values (B88). Also modified FractureDipSet.calculateTotalMacrofracturePopulation() and FractureDipSet.calculateCumulativeMacrofracturePopulationArrays() to calculate zero values and increments for P30 and P32 fracture densities if the fracture sets are not activated, and modified FractureDipSet.setMacrofractureDeactivationRate() to more efficiently calculate the mean active f
[…]
