### 2024-03-10 1a3726d
Modified PointXYZ.getCrossoverPoint()  to include a check for the minimum angle between the two lines; if it is less than the specified angular tolerance, the lines will be considered parallel and the function will return null (B84)
