### 2024-03-10 7a09ef7
Modified the PropagationControl object to assign a default value for the azimuth of fracture set zero which will be used if the load tensors for all deformation episodes are isotropic, and modified main() (for the standalone version) and ExecuteSimple() (for the Petrel plugin) to set this to the specified minimum horizontal strain azimuth for the first deformation episode. This prevents the fracture orienation defaulting to E-W if another minimum horizontal strain orientation is specified but the strain load is isotropic (B90).
Also added DFM_Plotter_6set_Centrelines.xlsm to data analysis spre
[…]
