+++ DFMGenerator_SharedCode/PropagationControl.cs
@@ -1221,6 +1221,10 @@ namespace DFMGenerator_SharedCode
             }
         }
         /// <summary>
+        /// Default azimuth of fracture set 0 - will be used if Applied_Epsilon_hmin_azimuth is not defined for any deformation episodes
+        /// </summary>
+        private double DefaultFractureAzimuth { get; set; }
+        /// <summary>
         /// Azimuth of minimum applied horizontal strain for the first deformation episode (radians) 
         /// </summary>
         public double Initial_Applied_Epsilon_hmin_azimuth
@@ -1228,14 +1232,14 @@ namespace DFMGenerator_SharedCode
             get
             {
                 // Look at each deformation episode in turn for a valid ehmin azimuth (i.e. an anisotropic stress or strain load)
-                foreach(DeformationEpisodeLoadControl deformationEpisode in deformationEpisodes)
+                foreach (DeformationEpisodeLoadControl deformationEpisode in deformationEpisodes)
                 {
                     double episode_hmin_azimuth = deformationEpisode.Applied_Epsilon_hmin_azimuth;
                     if (!double.IsNaN(episode_hmin_azimuth))
                         return episode_hmin_azimuth;
                 }
-                // If there are no deformation episodes with anisotropic stress or strain loads, return 0
-                return 0;
+                // If there are no deformation episodes with anisotropic stress or strain loads, return the specified default value, or zero if this is not specified
+                return DefaultFractureAzimuth;
             }
         }
 
@@ -1275,7 +1279,8 @@ namespace DFMGenerator_SharedCode
         /// <param name="timeUnits_in">Time units for deformation rates</param>
         /// <param name="CalculateFracturePorosity_in">Flag to calculate and output fracture porosity</param>
         /// <param name="FractureApertureControl_in">Flag to determine method used to determine fracture aperture - used in porosity and permeability calculation</param>
-        public void setPropagationControl(bool CalculatePopulationDistribution_in, int no_l_indexPoints_in, double max_HMin_l_indexPoint_Length_in, double max_HMax_l_indexPoint_Length_in, bool CalculateRelaxedStrainPartitioning_in, bool OutputBulkRockElasticTensors_in, StressDistribution StressDistribution_in, double max_TS_MFP33_increase_in, double historic_a_MFP33_termination_ratio_in, double active_total_MFP30_termination_ratio_in, double minimum_ClearZone_Volume_in, int maxTimesteps_in, double maxTimestepDuration_in, int no_r_bins_in, double minImplicitMicrofractureRadius_in, double FractureNucleationPosition_in, bool checkAlluFStressShadows_in, double anisotropyCutoff_in, bool WriteImplicitDataFiles_in, TimeUnits timeUnits_in, bool CalculateFracturePorosity_in, FractureApertureType FractureApertureControl_in)
+        /// <param name="DefaultFractureAzimuth_in">Default azimuth of fracture set 0 - will be used if Applied_Epsilon_hmin_azimuth is not defined for any deformation episodes</param>
[…]
