### 2024-04-17 #6178 @ranzhengcode
Dear patrick diehl,

I apologize for not having time to review the manuscript.

Zheng Ran

获取Outlook for Android<https://aka.ms/AAb9ysg>
