### 2024-06-28 #6178 @diehlpk
> Hi, sorry for the delay in replying. I'm not quite sure how this process works - do you need anything more from me right now? I cannot think of other potential reviewers at the moment but I could contact some of the reviewers I previously suggested to ask if they received the invitation to review, and if they are unable to review it themselves, if they could suggest other reviewers. With regard to making any required corrections, should I wait until the reviews are completed for this? I noticed that the auto-check picked up a broken DOI link - this should be very straightforward to fix.

No, I am still looking for a second reviewer. If you could mention potential reviewers that would be good.
