### 2024-07-26 #6178 @JointFlow
Hi, I have contacted Dave Healy directly to ask him again whether he would be able to review the paper (I had suggested him previously but I don't think he replied to the invitation). I will let you know as soon as he replies.

I have also updated the .bib file on the Github server to fix the broken doi link identified above. However I am not sure how to regenerate the pdf file and update it - are them some instructions on how to do this?
