### 2024-08-15 #6178 @DaveHealy-github
Hello
How do I:

  1.  View the ms
  2.  Review the ms
In this system?

Thanks
Dave

Prof. Dave Healy   [he/him/his]
Director of Geosolutions<http://www.geosolutions.leeds.ac.uk/>
Professor of Geomechanics
Staff profile page<https://environment.leeds.ac.uk/staff/12746/dave-healy>

Leading the energy transition through interdisciplinary subsurface research

My working day may not be your working day. Please do not feel obliged to reply to this e-mail outside of your normal working hours.

Room 8.107
Earth and Environment Building
School of Earth and Environment
University of Leeds
Leeds LS2 9JT

From: Patrick Diehl ***@***.***>
Sent: Wednesday, August 14, 2024 11:19 PM
Cc: David Healy ***@***.***>; Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [PRE REVIEW]: DFM Generator v2.2: a new method for simulating the growth of natural fracture networks (Issue #6178)

CAUTION: External Message. Use caution opening links and attachments.

Hi, I have spoken to Dave Healy again and he said that he would be happy to review the paper. His Github address is DaveHealy-github. Will you contact him to request a review?

I invited him.

-
