### 2024-10-22 #6818 @mohd-afeef-badri
Hello @diehlpk, @JointFlow,

First, I apologize for the delayed response, I have been swamped with other tasks. That said, I want to commend the authors for their efforts for the JOSS paper.

As part of my review, I intended to test the software, but I noticed that the repository only includes a Windows binary. Since I'm on Linux, I’m unable to run the binary. I tried to explore the option of compiling the code myself, but there are no Makefiles, CMake files, or other build instructions available.

@JointFlow, could you kindly provide guidance on how to compile the DFM Generator? Since this is an open-source project, it would be helpful to have clear instructions on how to compile the software included in the repository.
