### 2024-10-23 #6818 @JointFlow
Hi @mohd-afeef-badri,

The code is designed to be compiled in Visual Studio but the standalone version should compile in any C Sharp compiler, as it only uses standard C Sharp code. This version uses the C Sharp source files from DFMGenerator_Standalone and DFMGenerator_SharedCode. I am not familiar with other C Sharp compilers and development environments so I don't know the details of how to do this. The Petrel Plug-in version will only compile in Visual Studio on Windows, as it needs the Ocean SDK which is only available for those platforms.

Which compiler are you trying to use?
