### 2024-10-23 a8375d1
Updated version number to v2.4.0
