+++ DFMGenerator_SharedCode/Grid.cs
@@ -15,7 +15,7 @@ namespace DFMGenerator_SharedCode
         /// <summary>
         /// Program name and version number (hard coded)
         /// </summary>
-        public static string VersionNumber { get { return "DFMGenerator v2.2.3"; } }
+        public static string VersionNumber { get { return "DFMGenerator v2.4.0"; } }
 
         // Grid data
         /// <summary>
