### 2024-11-11 d38c776
Working on ME59 (do not populate cells with no fractures), ME60 (output fracture reactivation potential), ME61 (output fracture permeability tensor) and ME 62 (use present day stress to calculate fracture aperture). Work in progress.
