+++ DFMGenerator_Ocean/DFMGeneratorWorkstep.cs
@@ -955,8 +955,6 @@ namespace DFMGenerator_Ocean
                     double DepthAtDeformation = -1;
                     if (!double.IsNaN(arguments.Argument_DepthAtDeformation_default))
                         DepthAtDeformation = arguments.Argument_DepthAtDeformation_default;
-                    if (!double.IsNaN(arguments.Argument_YoungsMod_default))
-                        YoungsMod = arguments.Argument_YoungsMod_default; // Can also be set from grid property
                     Property DepthAtDeformation_grid = arguments.Argument_DepthAtDeformation;
                     if ((DepthAtDeformation_grid != null) && (DepthAtDeformation_grid.Grid != PetrelGrid))
                     {
@@ -1009,6 +1007,8 @@ namespace DFMGenerator_Ocean
                     IntermediateOutputInterval IntermediateOutputIntervalControl = (IntermediateOutputInterval)arguments.Argument_IntermediateOutputIntervalControl;
                     // Flag to output the macrofracture centrepoints as a polyline, in addition to the macrofracture cornerpoints
                     bool OutputCentrepoints = arguments.Argument_OutputCentrepoints;
+                    // Flag to calculate and output the fracture reactivation potential: this represents the fracture driving stress if positive, and the cohesionless distance to failure if negative
+                    bool CalculateFractureReactivationPotential = arguments.Argument_CalculateFractureReactivationPotential;
                     // Flag to output the bulk rock compliance and stiffness tensors
                     bool OutputBulkRockElasticTensors = arguments.Argument_CalculateBulkRockElasticTensors;
                     // Fracture connectivity and anisotropy index control parameters
@@ -1016,6 +1016,12 @@ namespace DFMGenerator_Ocean
                     bool CalculateFractureConnectivityAnisotropy = arguments.Argument_CalculateFractureConnectivityAnisotropy;
                     // Flag to calculate and output fracture porosity
                     bool CalculateFracturePorosity = arguments.Argument_CalculateFracturePorosity;
+                    // Flag to calculate and output fracture permeability tensors
+                    bool CalculateFracturePermeabilityTensor = arguments.Argument_CalculateFracturePermeabilityTensor;
+                    // Algorithm to use for calculating fracture permeability
+                    PermeabilityCalculationAlgorithm PermeabilityAlgorithm = (PermeabilityCalculationAlgorithm) arguments.Argument_PermeabilityAlgorithm;
+                    // Fracture types included in the fracture permeability tensor: Microfractures only; Macrofractures only; All fractures
+                    FractureType FractureTypesInPermeabilityTensor = (FractureType)arguments.Argument_FractureTypesInPermeabilityTensor;
                     // Implicit fracture population distribution functions will only be calculated if the implicit data is written to file
                     bool CalculatePopulationDistribution = WriteImplicitDataFiles;
                     // Number of macrofracture length values to calculate for each of the implicit fracture population distribution functions
@@ -1024,6 +1030,9 @@ namespace DFMGenerator_Ocean
                     // Set these values to the approximate maximum length of fractures generated, or 0 if this is not known (0 will default to maximum potential length - but may be much greater than actual maximum length)
                     double MaxHMinLength = 0;
                     double MaxHMaxLength = 0;
+                    // Flag to populate implicit fracture data in gridblocks with no fractures?
[…]
