+++ DFMGenerator_SharedCode/GridblockConfiguration.cs
@@ -31,6 +31,14 @@ namespace DFMGenerator_SharedCode
     /// Enumerator for return codes for the PropagateDFN function: 0 if the calculation runs to completion without errors; 1 if the gridblock geometry is not correctly defined; 2 if there is an error in the driving stress or propagation distance; 3 if fracture limit is hit
     /// </summary>
     public enum PropagateDFNReturnCode { Completed, GridblockGeometryError, DrivingStressError, NewFractureLimitExceeded }
+    /// <summary>
+    /// Enumerator for algorithms to calculate fracture permeability
+    /// </summary>
+    public enum PermeabilityCalculationAlgorithm { Oda1985 }
+    /// <summary>
+    /// Enumerator for the data used to calculate present day effective stress
+    /// </summary>
+    public enum PresentDayStressFrom { Strain, EffectiveStress, AbsoluteStress }
 
     /// <summary>
     /// Class representing an entire gridblock
@@ -1188,6 +1196,153 @@ namespace DFMGenerator_SharedCode
         /// </summary>
         public double Hmin_azimuth { get { return PropControl.Initial_Applied_Epsilon_hmin_azimuth; } }
 
+        // Present day effective stress - to use for calculating fracture aperture and permeability
+        /// <summary>
+        /// Present day Terzaghi effective stress tensor - can be used to override the stress at the time of deformation, to calculate present day fracture aperture and permeability
+        /// </summary>
+        public Tensor2S PresentDayStress { get; private set; }
+        /// <summary>
+        /// Flag to use present day effective stress tensor, instead of stress at the time of deformation, to calculate fracture aperture and permeability
+        /// </summary>
+        public bool UsePresentDayStress { get { return !(PresentDayStress is null); } }
+        /// <summary>
+        /// Specify the present day Terzaghi effective stress, to be used when calculating fracture aperture and permeability
+        /// </summary>
+        /// <param name="EffectiveStress">Tensor2S object defining the present day Terzaghi effective stress</param>
+        public void SetPresentDayStress(Tensor2S EffectiveStress)
+        {
+            PresentDayStress = EffectiveStress;
+            UpdatePresentDayStressOnFractures();
+        }
+        /// <summary>
+        /// Specify the present day Terzaghi effective stress, to be used when calculating fracture aperture and permeability
+        /// </summary>
+        /// <param name="SigmaEffXX">XX component of Terzaghi effective stress tensor (Pa)</param>
+        /// <param name="SigmaEffYY">YY component of Terzaghi effective stress tensor (Pa)</param>
+        /// <param name="SigmaEffZZ">ZZ component of Terzaghi effective stress tensor (Pa)</param>
+        /// <param name="SigmaEffXY">XY component of Terzaghi effective stress tensor (Pa)</param>
+        /// <param name="SigmaEffYZ">YZ component of Terzaghi effective stress tensor (Pa)</param>
+        /// <param name="SigmaEffZX">ZX component of Terzaghi effective stress tensor (Pa)</param>
+        public void SetPresentDayStress(double SigmaEffXX, double SigmaEffYY, double SigmaEffZZ, double SigmaEffXY, double SigmaEffYZ, double SigmaEffZX)
+        {
+            SetPresentDayStress(new Tensor2S(SigmaEffXX, SigmaEffYY, SigmaEffZZ, SigmaEffXY, SigmaEffYZ, SigmaEffZX));
+        }
+        /// <summary>
+        /// Specify the present day absolute stress and fluid pressure, to calculate an effective stress tensor to be used when calculating fracture aperture and permeability
+        /// </summary>
+        /// <param name="SigmaXX">XX component of absolute stress tensor (Pa)</param>
+        /// <param name="SigmaYY">YY component of absolute stress tensor (Pa)</param>
[…]
