### 2024-11-11 516aef6
Further work on ME59, ME60, ME61 and ME62. Still need to implement setting the present day stress in the Petrel version, and check for bugs in calculting permeability and fracture reactivation potential.
