+++ DFMGenerator_Ocean/DFMGeneratorWorkstep.cs
@@ -1494,11 +1494,23 @@ namespace DFMGenerator_Ocean
                         convertFromGeneral_AppliedTemperatureChange_list.Add(UseGridFor_AppliedTemperatureChange_list[deformationEpisodeNo] ? AppliedTemperatureChange_grid_list[deformationEpisodeNo].Template.Equals(GeneralTemplate) : false);
                         convertFromGeneral_AppliedUpliftRate_list.Add(UseGridFor_AppliedUpliftRate_list[deformationEpisodeNo] ? AppliedUpliftRate_grid_list[deformationEpisodeNo].Template.Equals(GeneralTemplate) : false);
                     }
+                    // Present day stress
+                    IUnitConverter toSIStressUnits = PetrelUnitSystem.GetConverterFromUI(StressTemplate);
+                    bool convertFromGeneral_EhminAzi_PresentDay = (UseGridFor_EhminAzi_PresentDay ? EhminAzi_PresentDay_grid.Template.Equals(GeneralTemplate) : false);
+                    bool convertFromGeneral_AppliedOverpressure_PresentDay = (UseGridFor_AppliedOverpressure_PresentDay ? AppliedOverpressure_PresentDay_grid.Template.Equals(GeneralTemplate) : false);
+                    bool convertFromGeneral_Sxx_PresentDay = (UseGridFor_Sxx_PresentDay ? Sxx_PresentDay_grid.Template.Equals(GeneralTemplate) : false);
+                    bool convertFromGeneral_Syy_PresentDay = (UseGridFor_Syy_PresentDay ? Syy_PresentDay_grid.Template.Equals(GeneralTemplate) : false);
+                    bool convertFromGeneral_Szz_PresentDay = (UseGridFor_Szz_PresentDay ? Szz_PresentDay_grid.Template.Equals(GeneralTemplate) : false);
+                    bool convertFromGeneral_Sxy_PresentDay = (UseGridFor_Sxy_PresentDay ? Sxy_PresentDay_grid.Template.Equals(GeneralTemplate) : false);
+                    bool convertFromGeneral_Syz_PresentDay = (UseGridFor_Syz_PresentDay ? Syz_PresentDay_grid.Template.Equals(GeneralTemplate) : false);
+                    bool convertFromGeneral_Szx_PresentDay = (UseGridFor_Szx_PresentDay ? Szx_PresentDay_grid.Template.Equals(GeneralTemplate) : false);
+                    bool convertFromGeneral_FluidPressure_PresentDay = (UseGridFor_FluidPressure_PresentDay ? FluidPressure_PresentDay_grid.Template.Equals(GeneralTemplate) : false);
                     // Depth at start of deformation
                     bool convertFromGeneral_DepthAtDeformation = (UseGridFor_DepthAtDeformation ? DepthAtDeformation_grid.Template.Equals(GeneralTemplate) : false);
                     // Young's Modulus
                     IUnitConverter toSIYoungsModUnits = PetrelUnitSystem.GetConverterFromUI(YoungsModTemplate);
                     bool convertFromGeneral_YoungsMod = (UseGridFor_YoungsMod ? YoungsMod_grid.Template.Equals(GeneralTemplate) : false);
+                    bool convertFromGeneral_YoungsMod_PresentDay = (UseGridFor_YoungsMod_PresentDay ? YoungsMod_PresentDay_grid.Template.Equals(GeneralTemplate) : false);
                     // Porosity
                     IUnitConverter toSIPorosityUnits = PetrelUnitSystem.GetConverterFromUI(PorosityTemplate);
                     bool convertFromGeneral_Porosity = (UseGridFor_Porosity ? Porosity_grid.Template.Equals(GeneralTemplate) : false);
[…]
