### 2024-11-18 #6818 @JointFlow
Hi @diehlpk I will try to see if I can find some other reviewers. The reason the software is Windows only is that the platform it uses (Petrel) is Windows only, so anyone who uses that platform should be on Windows. The code for the standalone version of the software should compile on other platforms as it has no Windows-specific components, although I have not tried this.
