### 2024-11-18 60779eb
Fixed bug B91 (incorrectly calculating initial horizontal stress when Biot's coefficient was <1) by including the effect of differential compaction of the grains and matrix in the calculation. The code was streamlined by moving the calculation of the equivalent initial stress relaxation from the frontends (Program.cs and DFMGeneratorWorkstep.cs) to a new function SetCricitcalInitialStressStrainState in the StressStrainState object. Tested but needs confirmation.
