+++ DFMGenerator_Standalone/Program.cs
@@ -537,7 +537,7 @@ namespace DFMGenerator_Standalone
             ModelTimeUnits = TimeUnits.ma;
             DeformationEpisodeDuration_list.Add(DeformationEpisodeDuration);*/
             // Add a deformation episode with uniaxial extension of -0.01/ma over 1ma
-            EhminAzi_list.Add(EhminAzi);
+            /*EhminAzi_list.Add(EhminAzi);
             EhminRate_list.Add(-0.01);
             EhmaxRate_list.Add(EhmaxRate);
             AppliedOverpressureRate_list.Add(AppliedOverpressureRate);
@@ -547,55 +547,56 @@ namespace DFMGenerator_Standalone
             DeformationEpisodeDuration_list.Add(1);
             AbsoluteStressRate_list.Add(AbsoluteStressRate);
             InitialFluidPressure_list.Add(InitialFluidPressure);
+            InitialAbsoluteStress_list.Add(InitialAbsoluteStress);*/
+            /*// Add an uplift episode, with uplift of 1800m over 18ma
+            EhminAzi_list.Add(EhminAzi);
+            EhminRate_list.Add(EhminRate);
+            EhmaxRate_list.Add(EhmaxRate);
+            AppliedOverpressureRate_list.Add(AppliedOverpressureRate);
+            AppliedTemperatureChange_list.Add(AppliedTemperatureChange);
+            AppliedUpliftRate_list.Add(100);
+            StressArchingFactor_list.Add(StressArchingFactor);
+            DeformationEpisodeDuration_list.Add(18);
+            AbsoluteStressRate_list.Add(AbsoluteStressRate);
+            InitialFluidPressure_list.Add(InitialFluidPressure);
+            InitialAbsoluteStress_list.Add(InitialAbsoluteStress);*/
+            // Add an overpressure and cooling episode (e.g. injection of cold fluid) for 10 years, with stress arching
+            EhminAzi_list.Add(EhminAzi);
+            EhminRate_list.Add(EhminRate);
+            EhmaxRate_list.Add(EhmaxRate);
+            AppliedOverpressureRate_list.Add(1E+12);
+            //AppliedTemperatureChange_list.Add(-5E+6);
+            AppliedTemperatureChange_list.Add(0);
+            AppliedUpliftRate_list.Add(AppliedUpliftRate);
+            StressArchingFactor_list.Add(1);
+            DeformationEpisodeDuration_list.Add(1E-5);
+            AbsoluteStressRate_list.Add(AbsoluteStressRate);
+            InitialFluidPressure_list.Add(InitialFluidPressure);
             InitialAbsoluteStress_list.Add(InitialAbsoluteStress);
-             /*// Add an uplift episode, with uplift of 1800m over 18ma
-             EhminAzi_list.Add(EhminAzi);
-             EhminRate_list.Add(EhminRate);
-             EhmaxRate_list.Add(EhmaxRate);
-             AppliedOverpressureRate_list.Add(AppliedOverpressureRate);
-             AppliedTemperatureChange_list.Add(AppliedTemperatureChange);
-             AppliedUpliftRate_list.Add(100);
-             StressArchingFactor_list.Add(StressArchingFactor);
-             DeformationEpisodeDuration_list.Add(18);
-             AbsoluteStressRate_list.Add(AbsoluteStressRate);
-             InitialFluidPressure_list.Add(InitialFluidPressure);
-             InitialAbsoluteStess_list.Add(InitialAbsoluteStress);*/
-             /*// Add an overpressure and cooling episode (e.g. injection of cold fluid) for 10 years, with stress arching
-             EhminAzi_list.Add(EhminAzi);
[…]
