+++ DFMGenerator_SharedCode/GridblockConfiguration.cs
@@ -34,7 +34,7 @@ namespace DFMGenerator_SharedCode
     /// <summary>
     /// Enumerator for algorithms to calculate fracture permeability
     /// </summary>
-    public enum PermeabilityCalculationAlgorithm { Oda1985 }
+    public enum PermeabilityCalculationAlgorithm { Oda1986, OdaCorrected1987 }
     /// <summary>
     /// Enumerator for the data used to calculate present day effective stress
     /// </summary>
@@ -2353,6 +2353,38 @@ namespace DFMGenerator_SharedCode
 
             return (TotalAllTips > 0 ? TotalConnectedTips / TotalAllTips : undefinedReturn);
         }
+        /// <summary>
+        /// Mean number of other macrofractures that each macrofracture is connected to - i.e. total number of connections (intersections or hard-linked relays) divided by total number of macrofractures
+        /// </summary>
+        /// <param name="ReturnNanForUndefined">Determine return value if there are no fractures: if true, will return Nan; if false, will return 0</param>
+        /// <returns>Ratio of (4 * sII_MFP30_total * sIJ_MFP30_total) / T_MFP30_total</returns>
+        public double ConnectionsPerMacrofracture(bool ReturnNanForUndefined)
+        {
+            double undefinedReturn = ReturnNanForUndefined ? double.NaN : 0;
+            double TotalConnections = 0;
+            double TotalFractures = 0;
+            foreach (Gridblock_FractureSet fs in FractureSets)
+            {
+                // Only hard-linked relays will be counted. These count as two intersections, one for each fracture segment
+                if (gd.DFNControl.LinkFracturesInStressShadow)
+                    TotalConnections += (2 * fs.combined_sII_MFP30_total());
+                // Intersections create two fracture connections, one on the terminating fracture and one on the terminated fracture
+                TotalConnections += (2 * fs.combined_sIJ_MFP30_total());
+                TotalFractures += (fs.combined_T_MFP30_total() / 2);
+            }
+
+            return (TotalFractures > 0 ? TotalConnections / TotalFractures : undefinedReturn);
+        }
+        /// <summary>
+        /// Function to get the fracture network permeability 
+        /// </summary>
+        /// <param name="connectionsPerFracture"></param>
+        /// <returns></returns>
+        public static double GetNetworkPermeabilityMultiplierFromConnections(double connectionsPerFracture)
+        {
+            double alpha = 0.17;
+            return 1 - Math.Exp(-alpha * connectionsPerFracture);
+        }
 
         // Functions to return fracture permeability tensor
         /// <summary>
@@ -2361,11 +2393,7 @@ namespace DFMGenerator_SharedCode
         /// <returns>Tensor2S object representing microfracture permeability</returns>
         public Tensor2S MicrofracturePermeability()
         {
-            Tensor2S microfracturePermeability = new Tensor2S();
-            foreach (Gridblock_FractureSet fs in FractureSets)
-                foreach (FractureDipSet fds in fs.FractureDipSets)
-                    microfracturePermeability += fds.Total_uF_Permeability();
-            return microfracturePermeability;
+            return MicrofracturePermeability(-1);
         }
         /// <summary>
         /// Permeability tensor for all current layer-bound macrofractures in the gridblock
@@ -2373,11 +2401,7 @@ namespace DFMGenerator_SharedCode
         /// <returns>Tensor2S object representing macrofracture permeability</returns>
         public Tensor2S MacrofracturePermeability()
         {
-            Tensor2S macrofracturePermeability = new Tensor2S();
-            foreach (Gridblock_FractureSet fs in FractureSets)
-                foreach (FractureDipSet fds in fs.FractureDipSets)
[…]
