+++ DFMGenerator_Ocean/DFMGeneratorWorkstep.cs
@@ -1925,8 +1925,11 @@ namespace DFMGenerator_Ocean
                         permeabilityLabel += " using the ";
                         switch (PermeabilityAlgorithm)
                         {
-                            case PermeabilityCalculationAlgorithm.Oda1985:
-                                permeabilityLabel += "Oda 1985";
+                            case PermeabilityCalculationAlgorithm.Oda1986:
+                                permeabilityLabel += "Oda (1986)";
+                                break;
+                            case PermeabilityCalculationAlgorithm.OdaCorrected1987:
+                                permeabilityLabel += "Oda corrected (1987)";
                                 break;
                             default:
                                 break;
@@ -4808,9 +4811,11 @@ namespace DFMGenerator_Ocean
                             Template LengthTemplate = PetrelProject.WellKnownTemplates.GeometricalGroup.Distance;
                             Template DeformationTimeTemplate = PetrelProject.WellKnownTemplates.PetroleumGroup.GeologicalTimescale;
                             Template ConnectivityTemplate = PetrelProject.WellKnownTemplates.MiscellaneousGroup.Fraction;
+                            Template ConnectionsPerFractureTemplate = PetrelProject.WellKnownTemplates.MiscellaneousGroup.General;
                             Template AnisotropyTemplate = PetrelProject.WellKnownTemplates.MiscellaneousGroup.General;
                             Template FracturePorosityTemplate = PetrelProject.WellKnownTemplates.PetrophysicalGroup.Porosity;
                             Template FractureReactivationPotentialTemplate = PetrelProject.WellKnownTemplates.GeomechanicGroup.StressTotal;
+                            Template SlipTendencyTemplate = PetrelProject.WellKnownTemplates.MiscellaneousGroup.Fraction;
                             Template StiffnessTensorComponentTemplate = PetrelProject.WellKnownTemplates.GeophysicalGroup.ModulusCompressional;
                             Template ComplianceTensorComponentTemplate = PetrelProject.WellKnownTemplates.GeophysicalGroup.CompressibilityRock;
                             Dictionary<Tensor2SComponents, Template> FracturePermeabilityTemplate = new Dictionary<Tensor2SComponents, Template>();
@@ -5149,12 +5154,16 @@ namespace DFMGenerator_Ocean
                                             if (CalculateFractureReactivationPotential)
                                             {
                                                 // Create properties and set templates for each property
-                                                Property MF_ReactivationPotential = FracSetData.CreateProperty(FractureReactivationPotentialTemplate);
-                                                MF_ReactivationPotential.Name = "Reactivation_Potential";
+                                                Property FDS_ReactivationPotential = FracSetData.CreateProperty(FractureReactivationPotentialTemplate);
+                                                Property FDS_SlipTendency = FracSetData.CreateProperty(FractureReactivationPotentialTemplate);
+                                                FDS_ReactivationPotential.Name = "Reactivation_Potential";
+                                                FDS_SlipTendency.Name = "Slip_Tendency";
 
                                                 // Add creation event to each property
-                                                IHistoryInfoEditor MF_ReactivationPotentialInfoEditor = HistoryService.GetHistoryInfoEditor(MF_ReactivationPotential);
-                                                MF_ReactivationPotentialInfoEditor.AddHistoryEntry(new HistoryEntry("Create dynamic implicit fracture model", "", PetrelSystem.VersionInfo.ToString()));
+                                                IHistoryInfoEditor FDS_ReactivationPotentialInfoEditor = HistoryService.GetHistoryInfoEditor(FDS_ReactivationPotential);
+                                                IHistoryInfoEditor FDS_SlipTendencyInfoEditor = HistoryService.GetHistoryInfoEditor(FDS_SlipTendency);
[…]

[…]
