### 2024-11-26 #6818 @diehlpk
Hi @T-Hageman do you have time to review this paper and have access to Windows?
