### 2024-11-26 #6818 @T-Hageman
> Hi @T-Hageman do you have time to review this paper and have access to Windows?

Dear @diehlpk , Unfortunately I won't be able to review the paper due to time constraints (note that to test the software and all its claimed functions fully, it is not just windows that is required but it is also linked to a commercial code "petrel" (to which I don't have access either))
