### 2024-11-26 #6818 @diehlpk
@JointFlow Can you might help to find reviewers? I think it will be very challenging to find reviewers with access to Windows and to the commercial petrel code. The JOSS reviewer list does not indicate the operating system and of course if the reviewer has access to petrel.
