### 2024-11-27 ac93649
Fixed error in calculation of Oda Anisotropy index; not yet completed
