+++ DFMGenerator_SharedCode/GridblockConfiguration.cs
@@ -2376,14 +2376,74 @@ namespace DFMGenerator_SharedCode
             return (TotalFractures > 0 ? TotalConnections / TotalFractures : undefinedReturn);
         }
         /// <summary>
-        /// Function to get the fracture network permeability 
+        /// Get the current fracture fabric tensor F, as defined by Oda 1983
         /// </summary>
-        /// <param name="connectionsPerFracture"></param>
         /// <returns></returns>
+        public Tensor2S FractureFabricTensor(FractureType fracType)
+        {
+            return FractureFabricTensor(fracType, -1);
+        }
+        /// <summary>
+        /// Get the fracture fabric tensor F, as defined by Oda 1983, at the end of a specified previous timestep
+        /// </summary>
+        /// <param name="Timestep_M">Index number of the specified timestep</param>
+        /// <returns></returns>
+        public Tensor2S FractureFabricTensor(FractureType fracType, int Timestep_M)
+        {
+            bool useCurrentDensityData = (Timestep_M < 0);
+            Tensor2S FTensor = new Tensor2S();
+            foreach (Gridblock_FractureSet fs in FractureSets)
+                foreach (FractureDipSet fds in fs.FractureDipSets)
+                {
+                    Tensor2S orientationTensor = fds.NormalVector ^ fds.NormalVector;
+                    double densityFactor = 0;
+                    if ((fracType == FractureType.Microfractures) || (fracType == FractureType.AllFractures))
+                        densityFactor += (3d / 16d) * (useCurrentDensityData ? fds.a_uFP33_total() + fds.s_uFP33_total() : fds.getTotaluFP33(Timestep_M));
+                    if ((fracType == FractureType.Microfractures) || (fracType == FractureType.AllFractures))
+                        densityFactor += (3d / 16d) * (useCurrentDensityData ? fds.a_MFP33_total() + fds.s_MFP33_total() : fds.getTotalMFP33(Timestep_M));
+
+                    FTensor += (densityFactor * orientationTensor);
+                }
+
+            return FTensor;
+        }
+        /// <summary>
+        /// Get the trace and the anistropy of the fracture fabric tensor for all fractures, as defined by Oda (1986), at the end of a specified previous timestep. 
+        /// This is used to calculate the fracture network permeability mutliplier, as defined in Oda et al. (1987). 
+        /// NB The anisotropy here is calculated from the second invariant of the modified fracture fabric tensor, and is not the same as the anisotropy calculated from fracture density above
+        /// </summary>
+        /// <param name="Timestep_M">Index number of the specified timestep; set neagtive to get current values</param>
+        /// <param name="F0">Reference parameter for the trace of the fracture fabric tensor</param>
+        /// <param name="Anisotropy">Reference parameter for the anisotropy of the fracture fabric tensor</param>
+        private void GetF0andAnisotropy(int Timestep_M, out double F0, out double Anisotropy)
+        {
+            Tensor2S FTensor = FractureFabricTensor(FractureType.AllFractures, Timestep_M);
+            F0 = FTensor.Trace;
+            double F0over3 = F0 / 3;
+            Tensor2S Fdashed = FTensor - new Tensor2S(F0over3, F0over3, F0over3, 0, 0, 0);
+            Anisotropy = Math.Sqrt(6 * Math.Abs(Fdashed.SecondInvariant)) / F0;
+            return;
+        }
+        /// <summary>
+        /// Function to get the fracture network permeability based on the mean number of conenctions per fracture, defined using data from Oda et al. 1987
+        /// </summary>
+        /// <param name="connectionsPerFracture">Mean number of connections per fracture</param>
[…]
