### 2024-12-11 #6818 @diehlpk
@JointFlow I think we reached the point to withdraw the paper. We worked quite a while to find reviewers which have access to the proprietary software needed to test the software. Please let me know if you oppose by Friday.
