### 2024-12-11 #6818 @diehlpk
@openjournals/jose-eics I would recommend to withdraw the paper by Friday since we could not find reviewers. The software used the commercial code "Petrel subsurface software [0]". I could not find reviewers with access to this software and therefore the pool of reviewer is very limited. In addition, the software only works on Windows. Maybe @JointFlow could find reviewers and resubmit the paper if he has two reviewers with access to Petrel.

[0] https://www.slb.com/products-and-services/delivering-digital-at-scale/software/petrel-subsurface-software/petrel
