### 2024-12-13 c719327
Working on implementing new algorithm for fracture permeability correcting for fracture size and connectivity; still in progress
