+++ DFMGenerator_SharedCode/FractureDipSet.cs
@@ -357,6 +357,11 @@ namespace DFMGenerator_SharedCode
         /// <returns></returns>
         public double getTotalMFP30() { return CurrentFractureData.Total_MFP30_M; }
         /// <summary>
+        /// Return the P31 value for all microfractures, static and dynamic, during the current timestep
+        /// </summary>
+        /// <returns></returns>
+        public double getTotaluFP31() { return CurrentFractureData.Total_uFP31_M; }
+        /// <summary>
         /// Return the mean linear density of all microfractures, static and dynamic, during the current timestep
         /// </summary>
         /// <returns></returns>
@@ -371,6 +376,11 @@ namespace DFMGenerator_SharedCode
         /// </summary>
         /// <returns></returns>
         public double getTotaluFP33() { return CurrentFractureData.Total_uFP33_M; }
+        /// <summary>
+        /// Return the P34 value for all microfractures, static and dynamic, during the current timestep
+        /// </summary>
+        /// <returns></returns>
+        public double getTotaluFP34() { return CurrentFractureData.Total_uFP34_M; }
         /// <summary>
         /// Return the P35 value for all microfractures, static and dynamic, during the current timestep
         /// </summary>
@@ -627,6 +637,12 @@ namespace DFMGenerator_SharedCode
         /// <returns></returns>
         public double getTotalMFP30(int Timestep_M) { if (Timestep_M < 0) Timestep_M = gbc.CurrentExplicitTimestep; return PreviousFractureData.getTotal_MFP30_M(Timestep_M); }
         /// <summary>
+        /// Return the P31 value for all microfractures, static and dynamic, at the end of a specified previous timestep
+        /// </summary>
+        /// <param name="Timestep_M">Index number of the specified timestep; set to -1 to use the current timestep in the explicit fracture calculation</param>
+        /// <returns></returns>
+        public double getTotaluFP31(int Timestep_M) { if (Timestep_M < 0) Timestep_M = gbc.CurrentExplicitTimestep; return PreviousFractureData.getTotal_uFP31_M(Timestep_M); }
+        /// <summary>
         /// Return the mean linear density of all microfractures, static and dynamic, at the end of a specified previous timestep
         /// </summary>
         /// <param name="Timestep_M">Index number of the specified timestep; set to -1 to use the current timestep in the explicit fracture calculation</param>
@@ -644,6 +660,12 @@ namespace DFMGenerator_SharedCode
         /// <param name="Timestep_M">Index number of the specified timestep; set to -1 to use the current timestep in the explicit fracture calculation</param>
         /// <returns></returns>
         public double getTotaluFP33(int Timestep_M) { if (Timestep_M < 0) Timestep_M = gbc.CurrentExplicitTimestep; return PreviousFractureData.getTotal_uFP33_M(Timestep_M); }
+        /// <summary>
+        /// Return the P34 value for all microfractures, static and dynamic, at the end of a specified previous timestep
+        /// </summary>
+        /// <param name="Timestep_M">Index number of the specified timestep; set to -1 to use the current timestep in the explicit fracture calculation</param>
+        /// <returns></returns>
+        public double getTotaluFP34(int Timestep_M) { if (Timestep_M < 0) Timestep_M = gbc.CurrentExplicitTimestep; return PreviousFractureData.getTotal_uFP34_M(Timestep_M); }
         /// <summary>
         /// Return the P35 value for all microfractures, static and dynamic, at the end of a specified previous timestep
         /// </summary>
@@ -1223,6 +1245,16 @@ namespace DFMGenerator_SharedCode
         /// <returns></returns>
         public double s_uFP30_total() { return MicroFractures.s_P30_total; }
         /// <summary>
+        /// Total P31 value for active microfractures
+        /// </summary>
[…]
