### 2024-12-18 #6818 @kthyng
Thanks @danielskatz. @diehlpk I didn't see the initial ping because it looks like it went to JOSE instead of JOSS (but I did get your more recent ping and just looking at it now).
